pets-api/requirements.txt
Flask>=0.12,<0.13
flask-mongoengine>=0.9,<0.10
Flask-Script>=2.0
jsonschema>=2.6,<2.7
mongoengine>=0.14,<0.15
py-bcrypt>=0.4,<0.5            
pymongo>=3.5,<3.6
pytest
